Enhance Cantina functionality: add order item retrieval endpoint, update Pedido entity to use Double for total value, and improve home_cantina.html with order detail modal and data attributes.

+++ src/main/java/br/ifsul/tcc/pediu_ifoi/domain/entity/Pedido.java
@@ -34,7 +34,7 @@ public class Pedido {
     @OneToMany(cascade = CascadeType.ALL)
     private List<ItemPedido> itensPedido;
 
-    private float valorTotal;
+    private Double valorTotal;
 
     private Date dataPedido;
